move IMaprTest (extention for IWebTest) to its own package
